# docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.Exam:
    properties:
      createdAt:
        type: string
      date:
        description: Date of the exam
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: Description of the exam
        type: string
      exam_type:
        allOf:
        - $ref: '#/definitions/model.ExamType'
        description: Exam type
      exam_type_id:
        description: ID of the exam type
        type: integer
      id:
        type: integer
      max_points:
        description: Maximum points for the exam
        type: integer
      name:
        description: Name of the exam
        type: string
      scores:
        allOf:
        - $ref: '#/definitions/model.Score'
        description: Scores for the exam
      updatedAt:
        type: string
    type: object
  model.ExamType:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        description: Name is the name of the exam type.
        type: string
      updatedAt:
        type: string
    type: object
  model.Grade:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        description: Name is the name of the grade.
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.Score:
    properties:
      comment:
        description: Comment represents any additional comments about the score.
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      exam_id:
        description: ExamID represents the ID of the associated exam.
        type: integer
      grade:
        allOf:
        - $ref: '#/definitions/model.Grade'
        description: Grade represents the associated grade.
      grade_id:
        description: GradeID represents the ID of the associated grade.
        type: integer
      id:
        type: integer
      points:
        description: Points represents the number of points scored by the student.
        type: integer
      student_id:
        description: StudentID represents the ID of the student.
        type: integer
      updatedAt:
        type: string
    type: object
...
  routes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: patrykhegenberg@gmail.com
    name: API support
    url: http://swagger.io/support
  license:
    name: Apache 2.0
    url: http://apache.com/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: Notenverwaltung API
  version: 0.0.8
paths:
...
  /exams:
    get:
      consumes:
      - application/json
      description: get all exams from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Exam'
            type: array
        "500":
          description: Fehler beim Abraufen der Exams
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all exams
      tags:
      - exam
    post:
      consumes:
      - application/json
      description: create a new exam
      produces:
      - application/json
      responses:
        "203":
          description: Non-Authoritative Information
          schema:
            $ref: '#/definitions/model.Exam'
        "400":
          description: ungueltige anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des Exams
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create exam
      tags:
      - exam
  /exams/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing exam
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: ungueltige exam-id
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim loeschen des Exams
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete exam
      tags:
      - exam
    get:
      consumes:
      - application/json
      description: get all exams from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Exam'
        "404":
          description: Exam nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: ungueltige Exam-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one exams by id
      tags:
      - exam
    put:
      consumes:
      - application/json
      description: update an existing exam
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Exam'
        "400":
          description: ungueltige exam-id
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Exam nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim aktualisieren des Exams
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update exam
      tags:
      - exam
  /examtypes:
    get:
      consumes:
      - application/json
      description: get all exam types from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ExamType'
            type: array
        "500":
          description: Fehler beim Abrufen der ExamTypeen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get all exam types
      tags:
      - examtype
    post:
      consumes:
      - application/json
      description: create a new exam type
      parameters:
      - description: ExamType object that needs to be added
        in: body
        name: examType
        required: true
        schema:
          $ref: '#/definitions/model.ExamType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ExamType'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Erstellen des ExamTypeen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: create exam type
      tags:
      - examtype
  /examtypes/:id:
    delete:
      consumes:
      - application/json
      description: delete an existing exam type
      parameters:
      - description: ExamType ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Ungültige ExamTypeen-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Fehler beim Löschen des ExamTypeen
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: delete exam type
      tags:
      - examtype
    get:
      consumes:
      - application/json
      description: get one exam type from db by ID.
      parameters:
      - description: ExamType ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ExamType'
        "400":
          description: Ungültige ExamTypeen-ID
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: ExamType nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: get one exam type by id
      tags:
      - examtype
    put:
      consumes:
      - application/json
      description: update an existing exam type
      parameters:
      - description: ExamType ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated exam type object
        in: body
        name: examType
        required: true
        schema:
          $ref: '#/definitions/model.ExamType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ExamType'
        "400":
          description: Ungültige Anfrage
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: ExamType nicht gefunden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: update exam type
      tags:
      - examtype
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"